+++ geoextent/lib/handleVector.py
@@ -5,12 +5,15 @@ from . import helpfunctions as hf
 import re
 from osgeo import osr
 
+null_island = [0] * 4
 fileType = "application/shp"
-search = { "time":["(.)*timestamp(.)*", "(.)*datetime(.)*", "(.)*time(.)*", "date$","^date"]}
+search = {"time": ["(.)*timestamp(.)*", "(.)*datetime(.)*", "(.)*time(.)*", "date$", "^date","^begin"]}
 logger = logging.getLogger("geoextent")
+
 
 def get_handler_name():
     return "handleVector"
+
 
 def checkFileSupported(filepath):
     '''Checks whether it is valid vector file or not. \n
@@ -33,8 +36,8 @@ def checkFileSupported(filepath):
         logger.debug("File {} is NOT supported by HandleVector module".format(filepath))
         return False
 
-def getCRS(filepath):
 
+def getCRS(filepath):
     dataset = ogr.Open(filepath)
     layer = dataset.GetLayer()
 
@@ -52,60 +55,85 @@ def getTemporalExtent(filepath):
     input "path": type string, file path to vector file
     '''
 
-    global parsed_time
-    dataset = ogr.Open(filepath)
-    layer = dataset.GetLayer()
-    layerDefinition = layer.GetLayerDefn()
+    datasource = ogr.Open(filepath)
+    layer_count = datasource.GetLayerCount()
+    logger.debug("{} contains {} layers".format(filepath,layer_count))
+    datetime_list = []
+    for layer in datasource:
 
-    field_names = []
-    for i in range(layerDefinition.GetFieldCount()):
-        field_names.append(layerDefinition.GetFieldDefn(i).GetName())
+        logger.debug("{} : Extracting temporal extent from layer {} ".format(filepath, layer))
+        layerDefinition = layer.GetLayerDefn()
+        field_names = []
+        for i in range(layerDefinition.GetFieldCount()):
+            field_names.append(layerDefinition.GetFieldDefn(i).GetName())
 
-    match_list = []
-    for x in search["time"]:
-        for j in field_names:
+        match_list = []
+        for x in search["time"]:
             term = re.compile(x, re.IGNORECASE)
-            match = term.search(j)
-            if match is not None:
-                match_list.append(j)
+            for j in field_names:
+                match = term.search(j)
+                if match is not None:
+                    match_list.append(j)
+        logger.debug(match_list)
 
-    if len(match_list) == 0:
-        logger.debug('The file {} has no TemporalExtent'.format(filepath))
+        if len(match_list) == 0:
+            logger.debug("File:{} /Layer: {}: No matched fields for temporal extent".format(filepath,layer))
+            pass
+        else:
+            datetime_list = []
+            for time_feature in match_list:
+                time_list = []
+                logger.debug(time_feature)
+                for feat in layer:
+                    time = feat.GetField(time_feature)
+                    if time is not None:
+                        time_list.append(time)
+                layer.ResetReading()
+
+                if len(time_list) != 0:
+                    parsed_time = hf.date_parser(time_list)
+                    if parsed_time is not None:
+                        datetime_list.extend(parsed_time)
+                    else:
+                        logger.debug('File:{} /Layer: {}: : Matched temporal extent "{}" field do not have recognizable time format'.format(filepath,layer,time_feature))
+                        pass
+                else:
+                    logger.debug("File:{} / Layer: {}: No values found in {} fields.".format(filepath,layer, time_feature))
+                    pass
+
+        logger.debug(match_list)
+    if len(datetime_list) == 0:
+        logger.debug("File {} do not have recognizable temporal extent".format(filepath))
         return None
     else:
-        timelist = []
-        for time_feature in match_list:
-            for feat in layer:
-                time = feat.GetField(time_feature)
-                if time is not None:
-                    timelist.append(time)
-
-    if len(timelist) == 0:
-        logger.debug('The file {} has no TemporalExtent'.format(filepath))
-        return None
-    else:
-        parsed_time = hf.date_parser(timelist)
-
-    if parsed_time is None:
-        logger.debug('The file {} has no recognizable TemporalExtent'.format(filepath))
-        return None
-    else:
-        tbox = [min(parsed_time).strftime('%Y-%m-%d'), max(parsed_time).strftime('%Y-%m-%d')]
+        tbox = [min(datetime_list).strftime(hf.output_time_format), max(datetime_list).strftime(hf.output_time_format)]
 
     return tbox
 
-def getBoundingBox(filepath):
-    ''' extracts bounding box from vector file \n
-    input "path": type string, file path to vector \n
-    returns bounding box of the file: type list, length = 4
-    '''
-    dataset = ogr.Open(filepath)
-    layer = dataset.GetLayer()
-    try:
-        ext = layer.GetExtent()
-        bbox_list= [ext[0],ext[2],ext[1],ext[3]]
-    except:
-        logger.debug("File {} does not have features".format(filepath))
-        bbox_list = None
 
-    return bbox_list
+def getBoundingBox(filepath):
+    """ extracts bounding box from vector file \n
+    input "filepath": type string, file path to vector \n
+    returns bounding box of the file: type list, length = 4
+    """
+    datasource = ogr.Open(filepath)
+    geo_dict = {}
+    for layer in datasource:
+        layer_name = layer.GetDescription()
+        ext = layer.GetExtent()
+        bbox = [ext[0], ext[2], ext[1], ext[3]]
+
+        try:
+            crs = layer.GetSpatialRef().GetAttrValue("GEOGCS|AUTHORITY", 1)
+        except Exception:
+            crs = None
+
+        geo_dict[layer_name] = {"bbox": bbox, "crs": crs}
+
+        if bbox == null_island or crs is None:
+            del geo_dict[layer_name]["crs"]
+        logger.debug(geo_dict)
+
+    bbox_merge = hf.bbox_merge(geo_dict, filepath)
+
+    return bbox_merge